pyproject.toml
[project]
name = "tubebench"
version = "1.0.0"
description = "MCN급 YouTube 채널 분석 & AI 영상 기획 웹 앱"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "google-api-python-client>=2.118.0",
  "python-dotenv>=1.0.1",
  "anthropic>=0.34.0",
  "google-generativeai>=0.8.0",
  "pandas>=2.2.0",
]

[project.scripts]
tubebench = "uvicorn:main"

[dependency-groups]
dev = ["streamlit>=1.33.0", "plotly>=5.20.0"]
